+++ FileCacheCore/src/main/java/com/soybeany/system/cache/core/token/SecretKeyHolder.java
@@ -46,7 +46,7 @@ public abstract class SecretKeyHolder implements Serializable {
         /**
          * 失效时间(millis)
          */
-        public final int expiryMillis;
+        public final long expiryMillis;
 
         public static WithExpiry deserialize(String content) {
             try {
@@ -66,7 +66,7 @@ public abstract class SecretKeyHolder implements Serializable {
 
         public WithExpiry(SecretKeyHolder holder, long expiryMillis) {
             super(holder);
-            this.expiryMillis = (int) expiryMillis;
+            this.expiryMillis = expiryMillis;
         }
     }
 }
