+++ SpringBootModule/FileCacheServer/src/main/java/com/soybeany/system/cache/server/model/DataInfo.java
@@ -14,7 +14,7 @@ public class DataInfo {
     public String contentType;
     public Long contentLength;
     public String contentDisposition;
-    public Integer pTtl;
+    public Long pTtl;
     public String md5;
     public String exInfo;
